requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
dotmap==1.3.30
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
h11==0.14.0
h3==4.2.2
httpcore==1.0.6
httpx==0.27.2
idna==3.7
ipdb==0.13.13
ipython==8.26.0
jedi==0.19.1
jiter
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
mutagen==1.47.0
numpy==2.2.6
openai==2.8.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
timezonefinder==6.5.9
tomli==2.0.1
tqdm==4.67.0
traitlets==5.14.3
twilio==9.3.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
uvicorn[standard]==0.30.6
wcwidth==0.2.13
yarl==1.15.1
google-cloud-pubsub==2.23.0
google-cloud-tasks==2.16.5
google-cloud-firestore==2.19.0
